+++ test_app/CMakeLists.txt
@@ -1,7 +1,8 @@
 # This is the project CMakeLists.txt file for the test subproject
 cmake_minimum_required(VERSION 3.16)
 
-set(EXTRA_COMPONENT_DIRS "../components/ mocks")
+list(APPEND EXTRA_COMPONENT_DIRS "../components/")
+list(APPEND EXTRA_COMPONENT_DIRS "mocks")
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-project(evse_test)
+project(test_app)
