+++ test_app/main/CMakeLists.txt
@@ -4,4 +4,5 @@ idf_component_register(SRC_DIRS "."
                     INCLUDE_DIRS "."
                     PRIV_INCLUDE_DIRS "../mocks/peripherals/include"
                     REQUIRES cmock evse
-                    PRIV_REQUIRES nvs_flash)
+                    PRIV_REQUIRES nvs_flash
+                    WHOLE_ARCHIVE)
