+++ test_app/pytest.ini
@@ -1,11 +1,11 @@
 [pytest]
 # only the files with prefix `test_` would be recognized as pytest test scripts.
-python_files = pytest_*.py
+python_files = test_*.py
 
 # additional options for pytest
 addopts =
   -s
-  --embedded-services esp,idf
+  --embedded-services esp,idf,wokwi
   --tb short
 
 # log related
